dekaf v2.0.62: adapt to std::format()

+++ khttp2.cpp
@@ -612,7 +612,7 @@ int Session::OnFrameRecv (const void* frame)
 {
 	auto Frame = static_cast<const nghttp2_frame*>(frame);
 
-	kDebug(3, "[stream {}] frame type: {}, category {}", Frame->hd.stream_id, TranslateFrameType(Frame->hd.type), Frame->headers.cat);
+	kDebug(3, "[stream {}] frame type: {}, category {}", Frame->hd.stream_id, TranslateFrameType(Frame->hd.type), static_cast<uint16_t>(Frame->headers.cat));
 
 	if (Frame->hd.type == NGHTTP2_HEADERS && Frame->headers.cat == NGHTTP2_HCAT_RESPONSE)
 	{
